made Interactable script that will parent all interactables
