+++ Assets/Scripts/Interactable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1458b6ce143775f47a753c2a244d43b8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
