+++ Assets/Scripts/Interactable/Interactable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ba1c639f0558da94ebf9988936a2fb8e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
